Progetto/SARIMA/requirements.txt
Flask==3.0.0
numpy==1.26.3
pandas==2.1.4
pmdarima==2.0.4
Requests==2.31.0
scipy==1.11.4
statsmodels==0.14.1
pyOpenSSL==23.3.0
urllib3==1.26.18
psutil==5.9.5
matplotlib==3.7.1